better files organization and adds some programs about the tower of hanoi puzzle
