+++ src/DataStructures/fundamental/arrays/GameEntry.java
@@ -1,0 +1,27 @@
+package DataStructures.fundamental.arrays;
+
+public class GameEntry {
+    private String name;    // name of the person earning this score
+    private int score;      // the score value
+
+    /**
+     * Constructs a game entry with given parameters
+     */
+    public GameEntry(String name, int score) {
+        this.name = name;
+        this.score = score;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public int getScore() {
+        return score;
+    }
+
+    @Override
+    public String toString() {
+        return "(" + name + ", " + score + ")";
+    }
+}
