--- src/fundamental/linked_lists/circularly_linked_lists/Node.java
@@ -1,23 +1,0 @@
-package fundamental.linked_lists.circularly_linked_lists;
-
-public class Node<E> {
-    private E element;
-    private Node<E> next;
-
-    public Node(E element, Node<E> next) {
-        this.element = element;
-        this.next = next;
-    }
-
-    public E getElement() {
-        return element;
-    }
-
-    public Node<E> getNext() {
-        return next;
-    }
-
-    public void setNext(Node<E> next) {
-        this.next = next;
-    }
-}
